updated unsignedBigInteger to integer
This solves errors when migrating using mysql database.

Errors occurred: "errno 150: Foreign key constraint is incorrectly formed"

To correctly form the foreign key. set the table to match users.id table "integer"

+++ database/migrations/2023_02_21_143845_bind_twofaccounts_and_groups_to_users.php
@@ -15,7 +15,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('twofaccounts', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')
+            $table->integer('user_id')
                   ->after('id')
                   ->nullable();
 
@@ -23,7 +23,7 @@ return new class extends Migration
         });
 
         Schema::table('groups', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')
+            $table->integer('user_id')
                   ->after('id')
                   ->nullable();
